runtime: move Qwen3.6 MoE contracts
