+++ crates/runner/src/qwen36_moe/speculative.rs
@@ -55,173 +55,11 @@
 //!   check confirms our drafts match `vllm.spec_decode.eagle.propose`.
 #![allow(dead_code)]
 
-/// Outcome of greedy-speculative accept-prefix logic.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AcceptPrefixOutcome {
-    /// Drafts accepted in order (length 0..=K). The base verified
-    /// `accepted_drafts[i]` produces the same prediction the base
-    /// would have made on its own at position `p+1+i`.
-    pub accepted_drafts: Vec<u32>,
-    /// Token to commit AFTER the accepted prefix. If `accepted_drafts.len()
-    /// < drafts.len()`, this is the base's predicted "correction" at
-    /// the rejection position (replacing the rejected draft). If all K
-    /// drafts were accepted, this is the "bonus" token: the base's
-    /// prediction at position `p + K + 1`, computed by feeding the
-    /// last accepted draft.
-    pub corrected_token: u32,
-}
-
-impl AcceptPrefixOutcome {
-    /// Total tokens to commit this speculative step:
-    ///   `accepted_drafts.len() + 1` (always at least 1, since
-    ///   `corrected_token` is always present).
-    pub fn n_emit(&self) -> usize {
-        self.accepted_drafts.len() + 1
-    }
-
-    /// All tokens to commit this speculative step in order.
-    pub fn emitted(&self) -> Vec<u32> {
-        let mut out = self.accepted_drafts.clone();
-        out.push(self.corrected_token);
-        out
-    }
-}
-
-/// Compute the greedy speculative accept-prefix outcome.
-///
-/// Inputs:
-///   - `drafts`: K candidate tokens proposed by the MTP head, where
-///     `drafts[i]` is the prediction for position `p + 1 + i`.
-///   - `base_predictions`: K+1 greedy-argmax predictions from the base
-///     model, where:
-///       * `base_predictions[0]` is the base's prediction for
-///         position `p + 1` (computed when `T_p` was sampled — we
-///         already have it from the previous decode step).
-///       * `base_predictions[i]` for `i in 1..=K` is the base's
-///         prediction for position `p + 1 + i`, computed by feeding
-///         `drafts[i-1]` as input at position `p + i`.
-///
-/// Output: see [`AcceptPrefixOutcome`].
-///
-/// # Early-termination caveat
-///
-/// In a real driver the base predictions past the first mismatch are
-/// never computed (we stop the verify loop early to save GPU work).
-/// This pure helper accepts them anyway — it just ignores them past
-/// the rejection point. The caller can pass shorter slices when early
-/// termination has already occurred, as long as `base_predictions.len()
-/// >= number_of_predictions_actually_computed`.
-///
-/// # Panics
-///
-/// Panics if `base_predictions.len() < drafts.len() + 1` AND no early
-/// rejection occurred (i.e., we'd need the bonus prediction but it
-/// wasn't supplied). Use [`accept_prefix_greedy_partial`] when the
-/// caller has already terminated early.
-pub fn accept_prefix_greedy(drafts: &[u32], base_predictions: &[u32]) -> AcceptPrefixOutcome {
-    let k = drafts.len();
-    let mut accepted: Vec<u32> = Vec::with_capacity(k);
-    for i in 0..k {
-        if i >= base_predictions.len() {
-            panic!(
-                "accept_prefix_greedy: base_predictions has only {} entries \
-                 but {} are needed to verify draft {i}; the caller terminated \
-                 the verify loop early without recording a corrected token. \
-                 Use `accept_prefix_greedy_partial` for the early-termination \
-                 case.",
-                base_predictions.len(),
-                i + 1
-            );
-        }
-        let b = base_predictions[i];
-        let d = drafts[i];
-        if b != d {
-            // Rejection at index `i`. Discard d_i and the rest;
-            // commit b_i as the correction.
-            return AcceptPrefixOutcome {
-                accepted_drafts: accepted,
-                corrected_token: b,
-            };
-        }
-        accepted.push(d);
-    }
-    // All K drafts accepted. The bonus token is base_predictions[K],
-    // produced by feeding drafts[K-1] (or by feeding T_p if K==0; the
-    // K==0 case degenerates to "no speculation, take base_predictions[0]
-    // as the next token").
-    if base_predictions.len() <= k {
-        panic!(
-            "accept_prefix_greedy: all {k} drafts accepted but \
-             base_predictions has only {} entries — need at least {} \
-             (one extra for the bonus token at position p+K+1).",
-            base_predictions.len(),
-            k + 1
-        );
-    }
-    AcceptPrefixOutcome {
-        accepted_drafts: accepted,
-        corrected_token: base_predictions[k],
-    }
-}
-
-/// Like [`accept_prefix_greedy`] but tolerates a base-predictions
-/// slice that ends at the rejection point. Used when the driver
-/// terminates the verify loop early (the production fast path —
-/// every skipped base step is a kernel launch saved).
-///
-/// Behaviour:
-///   - If the loop walked through all `drafts.len()` indices without
-///     rejection, the caller MUST also include the bonus prediction
-///     (so `base_predictions.len() >= drafts.len() + 1`) — otherwise
-///     this panics, since we cannot fabricate a bonus token.
-///   - If rejection occurred at index `i`, `base_predictions.len()`
-///     should be exactly `i + 1` (driver stopped after computing the
-///     mismatching prediction). Anything ≥ `i + 1` works; extra
-///     entries are ignored.
-pub fn accept_prefix_greedy_partial(
-    drafts: &[u32],
-    base_predictions: &[u32],
-) -> AcceptPrefixOutcome {
-    let n = base_predictions.len().min(drafts.len());
-    for i in 0..n {
-        let b = base_predictions[i];
-        let d = drafts[i];
-        if b != d {
-            return AcceptPrefixOutcome {
-                accepted_drafts: drafts[..i].to_vec(),
-                corrected_token: b,
-            };
-        }
-    }
-    // No rejection within the supplied slice — full prefix accepted up
-    // to whatever the slice covered.
-    if n < drafts.len() {
-        // Driver terminated before covering all drafts but reported no
-        // rejection. That's a contract violation: with no rejection
-        // the loop must have gone all the way to the bonus.
-        panic!(
-            "accept_prefix_greedy_partial: base_predictions stopped at \
-             {n} entries (covering drafts[0..{n}]) without a rejection, \
-             but drafts has {} entries. The driver must run the verify \
-             loop either to first rejection (inclusive) or all the way \
-             through to the bonus prediction.",
-            drafts.len()
-        );
-    }
-    // n == drafts.len() — all drafts accepted; need the bonus.
-    if base_predictions.len() <= drafts.len() {
-        panic!(
-            "accept_prefix_greedy_partial: all {n} drafts accepted but \
-             base_predictions has only {n} entries — need at least \
-             {} (one extra for the bonus token at position p+K+1).",
-            n + 1
-        );
-    }
-    AcceptPrefixOutcome {
-        accepted_drafts: drafts.to_vec(),
-        corrected_token: base_predictions[drafts.len()],
-    }
-}
+pub(crate) use supersonic_runtime::qwen36_moe::speculative::SpeculativeStepResult;
+#[cfg(test)]
+use supersonic_runtime::qwen36_moe::speculative::{
+    accept_prefix_greedy, accept_prefix_greedy_partial,
+};
 
 // ============================================================================
 // Phase 6.3c: speculative-decode driver (orchestration only)
@@ -250,35 +88,6 @@ use gpu_hal::{GpuBuffer, ScalarType};
 
 use crate::qwen36_moe_mtp::{run_mtp_draft_chain, MtpChainScratch, MtpForwardScratch};
 use crate::qwen36_moe_types::{MtpLayerBuffers, MultiLayerGeom, PositionPair};
-
-/// Result of one speculative-decode step.
-#[derive(Debug, Clone)]
-pub struct SpeculativeStepResult {
-    /// Tokens to commit this step, in order. Always length 1..=K+1
-    /// (always at least one token: the corrected/bonus token).
-    pub emitted_tokens: Vec<u32>,
-    /// Number of MTP draft tokens proposed by this step. This is the
-    /// denominator for acceptance-rate telemetry; it may be zero near
-    /// the end of a short generation where only the corrected token
-    /// slot remains.
-    pub n_drafted: usize,
-    /// Number of MTP drafts accepted (0..=K). Exposed for stats /
-    /// `--emit-stage-timings` / acceptance-rate logging.
-    pub n_accepted: usize,
-    /// Number of base-model verify chains run by this step. Sequential
-    /// verify stops at first rejection, while batched verify always
-    /// runs K+1 chains.
-    pub base_steps: usize,
-    /// Extra base-model replay chains after batched partial acceptance.
-    /// Sequential verify mutates state in order and therefore leaves
-    /// this at zero.
-    pub replay_steps: usize,
-    /// `[hidden]` BF16 little-endian — the final-hidden bytes from
-    /// the LAST base decode step that ran during this iteration.
-    /// Becomes the next speculative step's `h_base` per the vLLM
-    /// recurrent equation.
-    pub final_hidden_bytes: Vec<u8>,
-}
 
 /// Run one speculative-decode iteration.
 ///
